Adds a map, graphics, characters and mouse control
Adds the classes.js file, where classes will be stored. Currently the
Character class is stored there. The utility.js file hosts functions which
could be useful for the other classes. Also adds images and their handling
in game.js. Adds mouse events and their handling. Allows the map to be
moved around with click-and-drag. Adds the first character, Bob.