Image management and Trees.
Adds the ImageRepository class, which manages images, so they don't have
to be loaded repeatedly for each instance of an object. Adds the Tree
class, descendant of GameObject. Trees can be alive and dead and can
fall(). Adds images for the variants of a tree and removes an unused image
for Grass. Also refactors the main game code, so the game now waits for
all images to load before starting the first draw cycle. Bob declaration
and tree1 declaration was also moved to this postLoad().