Characters chop nearest tree and better handling of game functions
Adds the GameFunction class, which holds a function available to a player.
Characters now have a reference to their parent level and search for the
nearest tree themselves. CharacterLists can now also call chopNearestTree
for all their characters. Adds some small modifications to the game loop,
so that it displays fps and doesn't try to catch up on intervals bigger
than 0.25s.