Level 1 and more Spellbook improvements
Level 1 is half complete and the Spellbook was improved even further with
deletes and deletion of lines. Level.showTutorial() is now able to break
text if it's too long. Tutorial text can also be formatted using <c> at
the beginning of the line for code. This version is still very buggy.