Spellbook fixes, game pausing in background and improved movement
Spellbook now actually works, even though it is still in a testing state.
Changed the eval function for the console input for a function object
created from the console input and a 'library' of spellbook functions,
which is then immediately applied. The game now pauses if it's in the
background so it doesn't eat all the resources. There were slight
improvements in the order system and the .move() command now moves by
tiles, rather than pixels.