Farming - initial version
Adds the first iteration of farming. Introduces the FarmingPatch and
PlantType classes. A FarmingPatch is a GameObject and has a PlantType. A
FarmingPatch keeps track of its own growth and a PlantType keeps track of
the different images and stages that the growth can take. Side note: the
update function now iterates through all gameObjects, rather than just
throught the characters.