ItemPiles - items now appear on the floor
When you drop an item an ItemPile is created as a GameObject. These
ItemPiles now about the level and have an update function run on them each
frame. The code for creating Redberries was moved into ItemPiles and their
update function. Also includes a slight refactoring of buildTypes into
BuildTypes.